Requisito 5 e refatoraçao da pagina
